+++ test/state_manager/get_mixin_state_test.dart
@@ -73,19 +73,6 @@ void main() {
 
     expect(find.text("Count2: 1"), findsOneWidget);
   });
-
-  // testWidgets(
-  //   "MixinBuilder with build null",
-  //   (tester) async {
-  //     expect(
-  //       () => MixinBuilder<Controller>(
-  //         init: Controller(),
-  //         builder: null,
-  //       ),
-  //       throwsAssertionError,
-  //     );
-  //   },
-  // );
 }
 
 class Controller extends GetxController {